Added kernel page and frame allocator

+++ Kernel/Sources/lib/includes/queue.h
@@ -223,7 +223,7 @@ OS_RETURN_E queue_push(queue_node_t* node, queue_t* queue);
  */
 OS_RETURN_E queue_push_prio(queue_node_t* node, 
                             queue_t* queue, 
-                            const uint32_t priority);
+                            const uintptr_t priority);
 
 /**
  * @brief Removes a node from a queue.
